requirements.txt
aiogram==3.6.0
APScheduler==3.10.4
SQLAlchemy==2.0.35
pydantic==2.8.2
python-dotenv==1.0.1
aiohttp==3.9.5
uvloop==0.19.0 ; sys_platform == 'linux'
